LayoutTests/fast/mediastream/RTCPeerConnection-localDescription-expected.txt

Tests RTCPeerConnection localDescription.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pc.setLocalDescription(sessionDescription, requestSucceeded1, requestFailed1); did not throw exception.
PASS requestSucceeded was called.
PASS pc.setLocalDescription(sessionDescription, requestSucceeded2, requestFailed2); did not throw exception.
PASS requestFailed was called.
PASS pc.localDescription.type is "offer"
PASS successfullyParsed is true

TEST COMPLETE
